kunit: test: move death test to its own KUnit config
KUnit's death test is not hermetic and can affect behavior of other
tests (http://b/116313114#comment8); that being said, there is not
really any way to make it hermetic without substantially changing how
KUnit works (b/116655541), and the test has proven very useful. Thus,
this commit moves it to its own Kconfig, so we can run other tests
either with it or without it.

Signed-off-by: Brendan Higgins <brendanhiggins@google.com>
Google-Bug-Id: 116313114
Change-Id: I23c6e36834cc56a67ea589c67003fe4f031cb6db