perf hist: Fix bogus profiles when filters are enabled
When a filtered column is not present in the sort order, the profile becomes
arbitrary broken. Filtered and non-filtered entries are collapsed together,
and the filtered-by field ends up with a random value (either from
a filtered or non-filtered entry). If we end up with filtered entry/value,
then the whole collapsed entry will be filtered out and will be missing
in the profile. If we end up with non-filtered entry/value, then the overhead
value will be wrongly larger (include some subset of filtered out samples).

This leads to very confusing profiles. The problem is hard to notice,
and if noticed hard to understand. If the filter is for a single value,
then it can be fixed by adding the corresponding field to the sort order
(provided user understood the problem). But if the filter is for multiple
values, it's impossible to fix b/c there is no concept of binary sorting
based on filter predicate (we want to group all non-filtered values in
one bucket, and all filtered values in another).

Examples of affected commands:
perf report --tid=123
perf report --sort overhead,symbol --comm=foo,bar

Fix this by considering filtered status as the highest priority
sort/collapse predicate.

As a side effect this effectively adds a new feature of showing profile
where several lines are combined based on arbitrary filtering predicate.
For example, showing symbols from binaries foo and bar combined together,
but not from other binaries; or showing combined overhead of several
particular threads.

Change-Id: I14eb049038f89b8a0e1347d1b9723b5932f8b754
Signed-off-by: Dmitry Vyukov <dvyukov@google.com>